Add NewMachineSetup.sh and Homebrew/
The shell script NewMachineSetup.sh sources the files in the Homebrew
directory, which conspire to put together a Brewfile, install
Homebrew|Linuxbrew if it isn't already installed, tap homebrew/bundle,
and then install the software specified in the Brewfile.

This will make setting up a new machine in the future much easier.

**NOTE**: It is currently untested on Linux.