Set :padline no in bootstrap.sh
When tangling via emacs --batch (as the bootstrap.sh script does), the
results of the tangled file has extra newlines, which do not result when
tangling from inside of Emacs with C-c C-v t. I'm not sure why this is,
but it leads to a dirty working directory after running the bootstrap.sh
script. This change avoids that problem.