Don't use global patterns for Git LFS
If I were starting fresh with repos that didn't already contain binary
files tracked by git, this *might* make sense (still probably not,
though). But in particular, since I'm not, it's best not to do this,
otherwise Git LFS will constantly complain about files that it should
be tracking but isn't.