Update config for macOS w/Apple Silicon & Doom Emacs
I've decided to rename this repo from config-files to conf, so some
things had to be changed to accommodate that. Moreover, I now have a
machine running macOS on Apple Silicon, which required a few changes.
Homebrew is now installed into /opt/homebrew for macOS machines running
on Apple Silicon, so changes had to be made in order to accommodate
that.

Finally, I'm now really all in on using Doom Emacs. For a while, I had
been maintaining a vanilla Emacs configuration alongside a Doom Emacs
configuration, but it's time to really just abandon the vanilla Emacs
configuration.